refactor(e2e): extract metadata tests
